test: adapt for new header
