+++ frontend/svelte/src/tests/routes/Neurons.spec.ts
@@ -45,6 +45,12 @@ describe("Neurons", () => {
         );
     });
 
+    it("should render title", () => {
+      const { getAllByText } = render(Neurons);
+
+      expect(getAllByText(en.navigation.neurons).length).toBeGreaterThan(0);
+    });
+
     it("should render content", () => {
       const { getByText } = render(Neurons);
